almost done refactoring optimizeFilter onto dimension interface
