+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/dimension/FlagFromTagDimensionConfig.java
@@ -13,6 +13,8 @@ import com.yahoo.bard.webservice.data.dimension.impl.MapSearchProvider;
 import com.yahoo.bard.webservice.druid.model.dimension.extractionfunction.ExtractionFunction;
 import com.yahoo.bard.webservice.druid.model.dimension.extractionfunction.TagExtractionFunctionFactory;
 import com.yahoo.bard.webservice.util.DimensionStoreKeyUtils;
+import com.yahoo.bard.webservice.web.DefaultFilterOperation;
+import com.yahoo.bard.webservice.web.FilterOperation;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -23,6 +25,7 @@ import java.util.HashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -34,17 +37,42 @@ import javax.validation.constraints.NotNull;
  */
 public class FlagFromTagDimensionConfig extends DefaultRegisteredLookupDimensionConfig {
 
-    public static final String NULL_VALUE = "";
     public static final String DEFAULT_TRUE_VALUE = "true";
     public static final String DEFAULT_FALSE_VALUE = "false";
 
-    // TODO in reality we only need a set of base extraction functions to extend and the name of the filtering dimension
-    private final String filteringDimensionApiName;
+    public static final Set<FilterOperation> DEFAULT_POSITIVE_OPS = Stream.of(
+            DefaultFilterOperation.in,
+            DefaultFilterOperation.startswith,
+            DefaultFilterOperation.contains,
+            DefaultFilterOperation.eq
+    ).collect(Collectors.collectingAndThen(
+            Collectors.toSet(),
+            Collections::unmodifiableSet
+    ));
 
-    private final String tagValue;
-    private final String trueValue;
-    private final String falseValue;
-    private final Map<String, DimensionRow> rowMap;
+    public static final Set<FilterOperation> DEFAULT_NEGATIVE_OPS = Collections.unmodifiableSet(
+            Collections.singleton(
+                    DefaultFilterOperation.notin
+            )
+    );
+
+    public static final FilterOperation DEFAULT_POSITIVE_INVERTED_FILTER_OPERATION = DefaultFilterOperation.eq;
+    public static final FilterOperation DEFAULT_NEGATIVE_INVERTED_FILTER_OPERATION = DefaultFilterOperation.notin;
+
+    // TODO in reality we only need a set of base extraction functions to extend and the name of the filtering dimension
+    public final String filteringDimensionApiName;
+
+    public final String tagValue;
+    public final String trueValue;
+    public final String falseValue;
+
+    public final Set<FilterOperation> positiveOps;
+    public final Set<FilterOperation> negativeOps;
+
+    public final FilterOperation positiveInvertedFilterOperation;
+    public final FilterOperation negativeInvertedFilterOperation;
+
+    public final Map<String, DimensionRow> rowMap;
 
     /**
      * Configuration for a dimension that uses the presence of a tag as the basis for a true/false flag dimension.
@@ -75,69 +103,22 @@ public class FlagFromTagDimensionConfig extends DefaultRegisteredLookupDimension
             @NotNull String filteringDimensionApiName,
             String tagValue,
             String trueValue,
-            String falseValue
+            String falseValue,
+            Set<FilterOperation> positiveOps,
+            Set<FilterOperation> negativeOps,
+            @NotNull FilterOperation positiveInvertedFilterOperation,
+            @NotNull FilterOperation negativeInvertedFilterOperation
     ) {
         super(apiName, physicalName, description, longName, category, fields, defaultDimensionFields, keyValueStore, searchProvider, registeredLookupExtractionFns);
         this.filteringDimensionApiName = filteringDimensionApiName;
         this.tagValue = tagValue;
         this.trueValue = trueValue;
         this.falseValue = falseValue;
+        this.positiveOps = Collections.unmodifiableSet(positiveOps);
+        this.negativeOps = Collections.unmodifiableSet(negativeOps);
+        this.positiveInvertedFilterOperation = positiveInvertedFilterOperation;
+        this.negativeInvertedFilterOperation = negativeInvertedFilterOperation;
         this.rowMap = Collections.unmodifiableMap(rowMap);
-    }
-
-//    /**
-//     * Configuration for a dimension that uses the presence of a tag as the basis for a true/false flag dimension.
-//     * Uses the default true and false dimension values.
-//     *
-//     * @param apiName  The API Name is the external, end-user-facing name for the dimension.
-//     * @param description  A description of the dimension and its meaning.
-//     * @param longName  The Long Name is the external, end-user-facing long  name for the dimension.
-//     * @param category  The Category is the external, end-user-facing category for the dimension.
-//     * @param filteringDimensionApiName  The name of the tag dimension used when filter-serializing
-//     * @param groupingBaseDimensionApiName  The name of the string based tag dimension used as the basis for group by
-//     * operations
-//     * @param tagValue  The value whose presence in the flag dimen
-//     */
-//    public FlagFromTagDimensionConfig(
-//            @NotNull DimensionName apiName,
-//            String description,
-//            String longName,
-//            String category,
-//            @NotNull String filteringDimensionApiName,
-//            @NotNull String groupingBaseDimensionApiName,
-//            String tagValue
-//    ) {
-//        this(
-//                apiName,
-//                description,
-//                longName,
-//                category,
-//                filteringDimensionApiName,
-//                groupingBaseDimensionApiName,
-//                tagValue,
-//                DEFAULT_TRUE_VALUE,
-//                DEFAULT_FALSE_VALUE
-//        );
-//    }
-
-    public String getFilteringDimensionApiName() {
-        return filteringDimensionApiName;
-    }
-
-    public String getTagValue() {
-        return tagValue;
-    }
-
-    public String getTrueValue() {
-        return trueValue;
-    }
-
-    public String getFalseValue() {
-        return falseValue;
-    }
-
-    public Map<String, DimensionRow> getRowMap() {
-        return rowMap;
     }
 
     // TODO dependencies on already created dimensions can be refactored out
@@ -154,7 +135,11 @@ public class FlagFromTagDimensionConfig extends DefaultRegisteredLookupDimension
             @NotNull String filteringDimensionApiName,
             String tagValue,
             String trueValue,
-            String falseValue
+            String falseValue,
+            Set<FilterOperation> positiveOps,
+            Set<FilterOperation> negativeOps,
+            @NotNull FilterOperation positiveInvertedFilterOperation,
+            @NotNull FilterOperation negativeInvertedFilterOperation
     ) {
         // build search provider
         DimensionField keyField = fields.stream()
@@ -223,7 +208,11 @@ public class FlagFromTagDimensionConfig extends DefaultRegisteredLookupDimension
                 filteringDimensionApiName,
                 tagValue,
                 trueValue,
-                falseValue
+                falseValue,
+                positiveOps,
+                negativeOps,
+                positiveInvertedFilterOperation,
+                negativeInvertedFilterOperation
         );
     }
 }
